# docs/openapi/openapi.json
{
    "components": {"schemas":{"http.errorResponse":{"properties":{"error":{"type":"string"}},"type":"object"},"http.healthResponse":{"properties":{"status":{"type":"string"}},"type":"object"},"http.logResponse":{"properties":{"error":{"type":"string"},"id":{"type":"string"},"nodes_count":{"type":"integer"},"ports_count":{"type":"integer"},"status":{"type":"string"},"uploaded_at":{"type":"string"}},"type":"object"},"http.nodeResponse":{"properties":{"desc":{"type":"string"},"guid":{"type":"string"},"id":{"type":"integer"},"log_id":{"type":"string"},"port_guid":{"type":"string"},"sharp_info":{"additionalProperties":{"type":"string"},"type":"object"},"switch_info":{"additionalProperties":{"type":"string"},"type":"object"},"system_image_guid":{"type":"string"},"system_info":{"additionalProperties":{"type":"string"},"type":"object"},"type":{"type":"string"}},"type":"object"},"http.parseRequest":{"properties":{"path":{"type":"string"}},"type":"object"},"http.parseResponse":{"properties":{"error":{"type":"string"},"log_id":{"type":"string"}},"type":"object"},"http.portResponse":{"properties":{"guid":{"type":"string"},"id":{"type":"integer"},"lid":{"type":"integer"},"link_speed_actv":{"type":"integer"},"link_width_actv":{"type":"integer"},"node_id":{"type":"integer"},"phy_state":{"type":"integer"},"port_num":{"type":"integer"},"raw":{"additionalProperties":{"type":"string"},"type":"object"},"state":{"type":"integer"}},"type":"object"},"http.portsResponse":{"properties":{"ports":{"items":{"$ref":"#/components/schemas/http.portResponse"},"type":"array","uniqueItems":false}},"type":"object"},"http.topologyEdge":{"properties":{"port_a_id":{"type":"integer"},"port_b_id":{"type":"integer"}},"type":"object"},"http.topologyNode":{"properties":{"desc":{"type":"string"},"guid":{"type":"string"},"id":{"type":"integer"},"log_id":{"type":"string"},"type":{"type":"string"}},"type":"object"},"http.topologyResponse":{"properties":{"edges":{"items":{"$ref":"#/components/schemas/http.topologyEdge"},"type":"array","uniqueItems":false},"nodes":{"items":{"$ref":"#/components/schemas/http.topologyNode"},"type":"array","uniqueItems":false},"ports":{"items":{"$ref":"#/components/schemas/http.portResponse"},"type":"array","uniqueItems":false}},"type":"object"}}},
    "info": {"description":"REST API for parsing InfiniBand fabric logs (ibdiagnet2) and querying topology.","title":"Log Parser API","version":"1.0"},
    "externalDocs": {"description":"","url":""},
    "paths": {"/api/v1/log/{log_id}":{"get":{"description":"Returns log status, upload time, nodes/ports counts and optional error.","parameters":[{"description":"Log UUID","in":"path","name":"log_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.logResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Internal Server Error"}},"summary":"Get log meta","tags":["log"]}},"/api/v1/node/{node_id}":{"get":{"description":"Returns node attributes and optional info blocks (switch_info / system_info / sharp_info).","parameters":[{"description":"Node id","in":"path","name":"node_id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.nodeResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Internal Server Error"}},"summary":"Get node details","tags":["node"]}},"/api/v1/parse":{"post":{"description":"Opens the zip archive (path relative to data/), parses InfiniBand topology and saves it.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.parseRequest"}}},"description":"Path to log archive (relative to data/)","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.parseResponse"}}},"description":"log_id of the successfully parsed log"},"400":{"content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/http.parseResponse"},{"$ref":"#/components/schemas/http.errorResponse"}]}}},"description":"validation error (bad body, empty path, path outside data/)"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"internal error"}},"summary":"Parse a log archive","tags":["parse"]}},"/api/v1/port/{node_id}":{"get":{"description":"Returns the list of ports for the given node, ordered by port number.","parameters":[{"description":"Node id","in":"path","name":"node_id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.portsResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Internal Server Error"}},"summary":"List node ports","tags":["port"]}},"/api/v1/topology/{log_id}":{"get":{"description":"Returns nodes, ports and connections (edges) for the given log.","parameters":[{"description":"Log UUID","in":"path","name":"log_id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.topologyResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Not Found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.errorResponse"}}},"description":"Internal Server Error"}},"summary":"Get topology","tags":["topology"]}},"/health":{"get":{"description":"Returns 200 if the database is reachable, 503 otherwise.","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.healthResponse"}}},"description":"OK"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/http.healthResponse"}}},"description":"Service Unavailable"}},"summary":"Health check","tags":["health"]}}},
    "openapi": "3.1.0",
    "servers": [
        {"url":"localhost:8080"}
    ]
}